+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v1.7.0 (unreleased)
+  - Updated comment syndication feed to pull all comments instead of just articles
   - Allow non-logged-in users to specify a 'username' when submitting a comment
   - Allow silent edit for comment moderators / admins when editing a copy - prevents the 'Edited by' line from displaying
   - Forum Plugin - Improvements on how anonymous (non-logged-in) info displays in topics